👤♻️ feat: 유저 정보 조회 추가 및 미사용 컴포넌트 삭제
