--- src/features/post-detail/hooks/index.ts
@@ -1,1 +1,0 @@
-export { usePostDetail } from "./usePostDetail"
