--- src/features/post-detail/ui/index.ts
@@ -1,1 +1,0 @@
-export { PostDetailModal } from "./PostDetailModal"
