--- src/features/post-management/ui/index.ts
@@ -1,1 +1,0 @@
-export { PostDialogs } from "./PostDialogs"
