--- src/features/posts/api/index.ts
@@ -1,3 +1,0 @@
-export { postsApi } from "./postApi"
-export { tagsApi } from "./tagsApi"
-export { userApi } from "./userApi"
