--- src/features/posts/api/tagsApi.ts
@@ -1,7 +1,0 @@
-import { http } from "@/shared/lib"
-
-import { Tag } from "../types"
-
-export const tagsApi = {
-  getTags: () => http.get<Tag[]>("/posts/tags"),
-}
