--- src/features/posts/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useGetPosts } from "./usePost"
-export { useGetTags } from "./useTags"
