--- src/features/posts/lib/index.ts
@@ -1,1 +1,0 @@
-export { postQueryKeys } from "./queryKeys"
