--- src/features/posts/types/index.ts
@@ -1,3 +1,0 @@
-export type { Post, PostSearchParams, PostsResponse, PostWithUser } from "./post"
-export type { Tag } from "./tag"
-export type { User, UsersResponse } from "./user"
