--- src/features/posts/types/tag.ts
@@ -1,6 +1,0 @@
-/**태그 정보 */
-export interface Tag {
-  name: string
-  slug: string
-  url: string
-}
