--- src/features/user-management/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useUser } from "./useUser"
