--- src/features/user-management/ui/index.ts
@@ -1,1 +1,0 @@
-export { UserModal } from "./UserModal"
